src/serve/requirements.txt
fastapi>=0.115.12,<0.116.0
uvicorn>=0.34.3,<0.35.0
pydantic
pymongo>=4.13.1,<5.0.0
dotenv>=0.9.9,<0.10.0
requests>=2.32.3,<3.0.0
numpy>=1.2.4,<2.0.0
pandas>=2.2.3,<3.0.0
scikit-learn==1.6.1
tensorflow>=2.19.0,<3.0.0
python-dateutil
pytz
mlflow>=2.22.0,<3.0.0
